Extract 'me' controller
