+++ controllers/homeController.js
@@ -3,7 +3,6 @@
 const _ = require('lodash');
 const mongoose = require('mongoose');
 const passport = require('passport');
-const profileLib = require('../lib/profile');
 const userLib = require('../lib/user');
 const helpers = require('../lib/helpers');
 const passthrough = helpers.passthrough;
@@ -140,97 +139,12 @@ function postSignup(req, res) {
   })
 }
 
-function viewMyProfile(req, res) {
-  res.render('me/profile', {profile: req.user.profile});
-}
-
-function editMyProfile(req, res) {
-  res.render('me/profile_edit', { profile: req.user.profile });
-}
-
-function validationMessages(err) {
-  const messages = [];
-
-  const reasons = {
-    required: '% is required',
-    // todo: handle other kinds of validation errors
-    unknown: '% is invalid'
-  };
-
-  for (var attr in err.errors) {
-    const error = err.errors[attr];
-    const name = _.startCase(attr);
-    const reason = reasons[error.kind] || reasons.unknown;
-    messages.push(reason.replace(/%/, _.startCase(attr)));
-  }
-
-  return messages;
-}
-
-function updateMyProfile(req, res, next) {
-  var profile = req.user.profile;
-
-  profileLib
-    .updateProfile(profile, req.body)
-    .then(() => res.redirect('/me'))
-    .catch((err) => {
-      if (err.name === 'ValidationError') {
-        res.render('me/profile_edit', {
-          profile: profile,
-          messages: validationMessages(err)
-        });
-      } else {
-        next(err);
-      }
-    });
-}
-
 function viewProfile(req, res) {
   const profileId = req.param('profileId');
   ProfileService.fetch(profileId)
     .then((profile) => res.render('profile/view', {profile: profile}) )
     .catch( curriedHandleError(req, res) );
 }
-
-function showMemberPay(req, res) {
-  res.render('me/pay', {});
-}
-
-function postMemberPay(req, res) {
-  const amount = req.body.amount;
-  req.session.cart = {
-    kind: 'membership'
-    , description: 'Membership Share Purchase'
-    , amount: amount
-    , successMethodName: 'handleMembershipPaymentSuccess'
-  };
-  res.redirect('/pay');
-}
-
-require('./paymentController').mapMethod('handleMembershipPaymentSuccess', handleMembershipPaymentSuccess);
-
-function handleMembershipPaymentSuccess(req, res) {
-  console.log('handlemembershipsuccess cart: ' + _.inspect(req.session.cart));
-  console.log('old payment total: ' + req.user.profile.membershipPayments);
-  const profile = req.user.profile;
-  const amount = Number(req.session.cart.amount);
-  profile.membershipPayments = Number(profile.membershipPayments); // be damn sure we have a number!
-  profile.membershipPayments += amount;
-  if (profile.membershipPayments >= 25 && profile.memberType !== Profile.MEMBERSHIP_TYPES.provisional) {
-    profile.memberType = Profile.MEMBERSHIP_TYPES.full;
-  }
-  profile.save()
-    .then((saved) => {
-      delete req.session.cart;
-      res.redirect('/me/thanks');
-    })
-    .catch(curriedHandleError(req, res));
-}
-
-function membershipThanks(req, res) {
-  res.render('me/thanks', {});
-}
-
 
 function logout(req, res) {
   req.logout();
@@ -258,13 +172,6 @@ function addRoutes(router) {
   router.get('/signup', showSignup);
   router.post('/signup', postSignup);
   router.get('/afterAuth', afterAuth);
-  router.get('/me', viewMyProfile);
-  router.get('/me/edit', editMyProfile);
-  router.post('/me/edit', updateMyProfile);
-  router.get('/me/pay', showMemberPay);
-  router.post('/me/pay', postMemberPay);
-  router.get('/me/thanks', membershipThanks);
-  router.get('/m/:profileId', viewProfile);
   router.get('/logout', logout);
 
   router.get('/dump', dump);
